improvement(map): have search boxes directly init with map-state
This will allow search boxes to be anywhere within the react
component hierarchy, and doesn't require them to live under the
map component
